requirements.txt
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.3
pymongo==3.2.2
python-slugify==1.2.0
requests==2.9.1
requests-oauthlib==0.6.1
Unidecode==0.4.19
Werkzeug==0.11.5
wheel==0.24.0
